CHASM: extract struct to separate files
